refactor(common/manager): useDeviceHasUpdatesAvailable to not use deprecated apis
